Refactor the main architecture again to be based on "features"
The main idea is that the `RstLanguageServer` class and related code
contained in `esbonio/lsp/__init__.py` manages the lifecycle and
boilerplate code of the server.

Features are then registered with the server in a manner heavily
inspired by the way Sphinx itself is architected. Through common
method naming conventions these features then tie into the events they
need and the server ensures each method is called when required