lsp: Add ability to set filepath in completion tests
Some test cases require the ability to set the filepath of the file that
completion suggestions are being requested for. This commit introduces
a new `filepath` keyword argument for this purpose.

Additionally the `expected` and `unexpected` arguments have also been
made into keyword-only arguments so that new arguments can be added in
the future without worrying about the order they are given in at call
sites.