lsp: Initial implementation of filepath completions.
This commit provides an initial implementation of filepath completions
for the `image`, `figure` ,`include`, `literalinclude` directives as
well as the `download` role.

Relative and absolute paths are handled, but no consideration is given
to the file types (e.g. more than just images are suggested to
`.. image::`). Toctree entry completions are also not currently
implemented.