lsp: Allow clients to override default conf.py discovery
Almost all initialization has been moved out of the `initialize` request
and into the `initialized` notification. This allows us to make a
request for the client's configuration and use it while constructing the
Sphinx application.

Clients can now provide the `esbonio.sphinx.confDir` option to bypass
the default `conf.py` discovery code and use a given absolute path. Or a
path may be given like `${workspaceRoot}/some/dir`, in which case the
path will be expanded relative to the workspace root as specified by the
language client during `initialize`