lsp: add ability to override Sphinx's srcdir
Clients can now use the `esbonio.sphinx.srcDir` option to override the
default assumption that Sphinx's `srcdir` should be the same as
`confdir`.

This can take the form of an absolute path or with the
`${workspaceRoot}` and `${confDir}` variables be a path relative to the
workspace root or Sphinx's confdir respectively.