vscode: Improve the development experience
This commit introduces the ability for the extension to connect to a
language server instance over TCP. This way of connecting is tried when
the `VSCODE_DEBUG` environment variable is present and set to `true`.

The commit also updates the `.vscode/launch.json` file to set this
environment variable as well as introducing a new launch configuration
that starts the language server in TCP mode.

Finally the inclusion of a compound configuration allows both the
language server and VSCode extension to be debugged simultaneously!

This implementation was heavily inspired from the example json-extension
in the pygls repo.
https://github.com/openlawlibrary/pygls/tree/e1a5c751fe45d940350d8a755490889b75d52273/examples/json-extension