vscode: Proof of concept preview.
This commit introduces a proof of concept implementation of a preview
for documentation built with the `html` builder.

Using `jsdom` we rewrite the URLs of any of the local assets (css, js,
img) etc. to the result of the `webview.asWebviewUri` method so that can
be loaded in the sandboxed environment.

It correctly determines the corresponding html file to the current
editor by inspecting the `SphinxConfig` returned from the language
server. However, it does not currently handle document reloads or
following the user as they move between source files.