lsp: Make use of `initializationOptions`
Originally we deferred most of the startup code to the `initialzed`
notification handler and used the `workspace/configuration` request to
obtain the values we needed to create a Sphinx application. However,
the `initialize` request has a field `initializationOptions` where the
client can pass values to the server on initialization.

This commit moves our startup code back into the `initialize` request
handler and looks for the relevant values as fields within
`initializationOptions`.

Additionally, the following cli arguments have been replaced with
configuration parameters in `initializationOptions`

- `--cache-dir` is now `esbonio.sphinx.buildDir`
- `--log-level` is now `esbonio.server.logLevel`
- `--log-filter` is now `esbonio.server.logFilter`
- `--hide-sphinx-output` is now `esbonio.server.hideSphinxOutput`