lsp: Reintroduce the `RstLanguageServer` class
The main idea behind the next series of commits is to have the
language server be the source of truth for all the information that
`LanguageFeature`s will query when processing whatever they need to
process.

The `RstLanguageServer` class still serves as the base that everything
will build on but it is now additionally responsible for providing
information on anything that is handled directly by `docutils`.
Currently this translates into indexing any role or directive
definitions that are registered directly with docutils.

Additionally this class is now responsible for managing all
diagnostics that are to be pushed to lanaguge clients.
`LanguageFeature`s that wish to provide diagnostics must now specify a
`source` as well as the `textDocument`'s URI.