lsp: Introduce `SphinxLanguageServer` class
The `SphinxLanguageServer` class builds on the `RstLanguageServer`
class in order to specialise in providing information about Sphinx
projects.

All the functionality that was the original `esbonio.lsp.sphinx`
module such as creating the Sphinx application and running builds is
now part of the server itself.

Build errors are still reported as Diagnostics but instead of trying
to parse stdout for errors, the `SphinxLogHandler` now intercepts log
messages allowing us to easily extract and report the full error
report.

The `SphinxLanguageServer` builds on the `get_roles` and
`get_directives` methods to include the roles and directives that
registered directly with Sphinx. It also provides additional methods
that can be used to query additional information such as Sphinx
domains. Instead of indexing all this information up front, it is now
done on demand and cached which should improve the startup time of the
server.

Finally, when it comes to testing all tests will be done by actually
running the server and communicating to it using the LSP protocol
itself. This should make the test cases much more representative and
give us the freedom to change implementation details without worrying
about updating the test cases.