lsp: Realign directive support
This commit realigns the `Directives` language feature to changes in
the language server.

Instead of indexing all directive definitions and their arguments, the
language feature now queries the server for these and can focus on
determining when it is appropriate to suggest completions etc.

On the subject of completions, the language feature now delegates to a
list of `ArgumentCompletion` providers when it's appropriate. This
makes it easy for new providers to be contributed without having to
worry about determining the context.