lsp: Realign role support
This commit realigns the `Roles` language feature to changes in the
language server.

Instead of indexing all role definitions and their targets, the
language feature now queries the server for them and can focus on
determining when it is appropriate to suggest completions etc.

On the subject of completions, the language feature now delegates to a
list of `TargetCompletion` providers when appropriate. This makes it
easy for new providers to be contributed without having to worry about
determining the context.