lsp: Proof of concept goto definition functionality
This adds a proof of concept implementation of the goto definition
feature of the language server protocol.

In a similar way to completions, language features can provide a list of
`definition_triggers` which determine when the `definition` method
should be called.

The `Roles` language feature has been extended to trigger a definition
lookup on complete role definitions which then delegates to a list of
`TargetDefinition` providers which do the actual lookup.

This currently only provides goto definition support for `:doc:` roles