lsp: Add `CompletionContext`
Since the number of arguments being associated with a completion request are
growing, it makes sense to introduce an object which can carry them all. In
addition to the existing `doc`, `match`, and `position` arguments, this commit
also introduces a new `location` argument.

This can be used to determine whether the completion request has been made

- from within an `*.rst` file
- from within a block of python code, or
- from within a docstring.