lsp: Send `initialized` notification as part of `ClientServer` startup
To align with the change where the initial Sphinx build has been moved out of the
handler for the `initialize` request and into the handler for the `initialized`
notification, it's now necessary for test cases to send this notification to the server.

Additionally, so that test cases can be sure that the server is in a good state to
continue, it's necessary for the test client to `await` on the receipt of the
`esbonio/buildComplete` notification before continuing.

Currently, as there is no built in mechanism within `pygls` to do this, the
`esbonio.lsp.testing` module includes a `ClientProtocol` class that builds on the
`LanguageServerProtocol` class to add `wait_for_notification` and
`wait_for_notification_async` methods. This extended protocol is then used when
constructing the `TestClient`, allowing for test cases to `await` on particular
notifications.