lsp: Specify `TextEdit` in role target completion items
The `complete_targets` method will now automatically fill out the
`text_edit` field on the `CompletionItems` it receives from the
providers so that they play nice with any existing text.

The `TextEdit` will be written in a way that completely replaces the
existing target with the target specified by the selected
`CompletionItem`. Both "plain" and "aliased" targets are handled.

However, if a provider wishes to override the default behavior it simply
has to fill out the `text_edit` field itself and the `complete_targets`
method will not interfere.

This also modifies the `PARTIAL_PLAIN_TARGET` and
`PARTIAL_ALIASED_TARGET` regular expressions so that they can also match
complete roles.