lsp: Align to the new default `TextEdit` behavior.
This modifies the `CompletionItems` that are generated to represent
intersphinx targets so they work well with the new `TextEdit` behavior
added to role target completions.

By setting `insert_text` to also include the target's project, the new
`TextEdit` does the right thing and we don't have to worry about doing
anything custom.