lsp: Override the default `TextEdit` for filepath completions
Filepath completions work by providing completions for a single segment
at a time. This means in a situation like the following:

   /path/to/some
            file.txt
	    folder/
	    test.png

When the user picks an item we don't want the default behavior to kick
in and replace the other path segments! So this commit updates the
`path_to_completion_item` function to generate a `TextEdit` that only
modifies the most recent part of the path (the string `some` in the
example above)