lsp: Change how `texDocument/completion` requests are handled.
Now that completion providers need to see the existing text in order to
provide accurate `CompletionItems` we now need to look at the entire
line in order to correctly process completions.

Additionally we handle the case where a pattern could match multiple
times on a single line by only generating suggestions when the matched
text contains the position at which the completion was requested.