lsp: Use a single regular expression for all role scenarios
Now that we need to be able to handle a much wider range of scenarios
when it comes to role based completions, having multiple regular
expressions to handle the different situations was getting unwieldy.
In particular it was getting harder to ensure the right expression was
being triggered for the situation.

So instead there is now a single regular expression that can handle
all stages of completion for a role and this commit aligns the rest of
the server to that setup.