testing: Attempt to resolve hanging tests
Add `pytest-timeout` as a dependency of the test suite. By adding a
timeout to the appropriate tests, if they hang in the future they should
fail and move on rather than timing out the build in CI

The underlying issue appears to be that we're sending the `initialize`
request before either the client or server were fully spun up, leaving
the `transport` attribute on the protocol object undefined. By adding a
short sleep if either `transports` are `None` we *should* prevent the
tests from hanging in the future.

I guess time will tell :)