code: Add "Switch/Pick Environment" options to pop ups
When the user has an environment configured that cannot be used to run
the language server for whatever reason, include an option to pick a
different environment using the Python extension's
`python.setInterpreter` command.

By building that selection opportunity into the setup workflow, the
extension can automatically retry running the language server reducing
some of the friction in getting a working environment setup.