docs: Add section on deployment
Add a "Deployment" section to the tutorial extension's documentation
outlining how generated tutorial files can be included in a Python
package.

Also update `esbonio-extensions` packaging to include the sample
tutorials from the documentation, as well as an
`python -m esbonio.tutorial` command that launches the sample
tutorials in a JupyterLab instance.