lsp: Add experimental spell checking module.
This adds experimental spell checking support with mispellings reported
as diagnostics and suggestions provided as code actions. Future commits
will provide a way for users to opt-in to this module, rather than it
be enabled by default.

It's experimental as the language is hardcoded to English and the error
detection/suggestions are fooled by certain constructs e.g. "quoted"
words, bulleted list and figure captions.