lsp: Introducing entry points and extensions.
This commit starts to expose the extensible way in which esbonio has
been written by extending the cli interface and introducing the idea of
entry points.

An entry point is a python module responsible for configuring and
running a language server instance. As of this commit there are now two
entry points built into the library.

- `esbonio.lsp.sphinx`: The Sphinx language server that's been available
  since the beginning
- `esbonio.lsp.rst`: A language server that works with the flavour of
  reStructuredText implemented by docutils.

To use a given entry point, set the server start command in your
language client of choice to `python -m <entry_point>`.

**Note:** the original `python -m esbonio` entry point remains and is an
alias for the default configuration (currently `esbonio.lsp.sphinx`)

Additionally, the standard cli interface for the bundlded language
servers has been extended with the ability to include/exclude additional
modules from the default configuration.