lsp: Add `completionItem/resolve` support for directives
Since directive implementations can be overriden, the name of a
directive is not a reliable way to identify them. So we now use
the `detail` field on a `CompletionItem` to carry the fully
qualified name of the class that implements them.

With some additional metadata in the `data` field on a
`CompletionItem`, the directives language feature now uses the
`completionItem/resolve` request to provide documentation for
a directive and its options.

This relies on the langauge feature being loaded up with the
documentation for the relevant items as it typically not available
as docstrings on the implementation. This is done through the
`add_documentation` method and by default the language server will
look for the following files.

- `esbonio/lsp/rst/directives.json`
- `esbonio/lsp/sphinx/directives.json`

Which should contain the documentation for `docutils` and `sphinx`
directives respectively.