script: Add script to scrape sphinx documentation
This script uses the `objects.inv` file published by the Sphinx
documentation site to discover and scrape the documentation for roles
and directives added by Sphinx. The resulting documentation is then
saved as a JSON file that can be loaded by the language server.

It is far from perfect but its probably good enough for now and we can
continue to refine it over time.

This commit also adds the result of running the script.