script: Add script to scrape docutils documentation
This adds the `scripts/generate_docutils_documentation.py` script which
pulls the latest rst source files for the reference directive and roles
documentation pages from the docutils docs site.

Example usage:

   python scripts/generate_docutils_documentation.py -o lib/esbonio/esbonio/lsp/rst/

The script produces a file `directives.json` with the following
structure

```
{
  "name(module.path.ClassName)": {
    "is_markdown": true,
    "description": [
      "# Some title",
      "This is useful when...",
      ...
    ]
    "options": {
      "opt1": "...",
      ...
    }
  }
}
```

As well as a file `roles.json` with a similar structure

This corresponds with the dictionary structure expected by the
`Directives` and `Roles` language features.

This commit also adds the result of running the script.