lsp: Add html preview to `SphinxLanguageServer`
This adds an implementation of the `esbonio.server.preview` command for
Sphinx projects that use the `html` builder. It works by standing up a
`http.server` pointed at the project's build directory in a background
process.

The command accepts a single options object but it can be omitted where
the default values will be used instead. Here is that options object and
its default values:

```json
{ "show": true }
```

For clients that do not provide their own preview, they can use the
default behaviour, where the server will issue a `window/showDocument`
request to open the preview in a web browser.

For clients that provide their own preview mechanism, they can suppress
this behavior by passing `{ "show": false }`. Instead they can use the
return value of the command `{"port": 1234}` which holds the port number
on `localhost` where the server is running.