code: Switch preview mechanism to an `<iframe>`
This drops the previous approach of trying to rewrite all urls to work
within a webview context and instead loads the content over a http
connection using a `<iframe>`.

By using an iframe, we retain the ability to switch between pages and
"follow" the user around as they navigate between documents.