lsp: Move to a new testing approach
Most of the contents of the `esbonio.lsp.testing` module has been moved
out into a new package and pytest plugin `pytest-lsp`.

The new plugin allows for "proper" end-to-end testing as it now runs the
server in a separate sub-process and interacts with it over the LSP
protocol - just as a regular client would do.

The new plugin also attempts to mimick different language clients by
adjusting the value of `ClientCapabilities` that is sent to the server
during initialization.

Finally the new plugin will attempt to help enforce that a server
implementation follows the spec by (eventually) providing checks that
ensures the server respects the capabilities broadcast by the client