lsp: Lookup `:ref:` targets by name
It turns out that when docutils transforms a label into the id used in
the parsed doctree, it also maintains a mapping of the "names" that
correspond to these ids.

To ensure that the language server correctly handles all forms of label
it now uses this map to look up the correct id to search for.