lsp: Add `get_initial_doctree` method
Buried in the Sphinx source is an *awesome* idea, using their
`CustomReSTDispatcher` class it's possible to override the mechanism
docutils uses to look up role and directive implementations.

Sphinx uses it to implement domains, but we can use it to replace
**all** roles and directives with our own dummy implementations. This
gives us the ability to construct a sort of AST for a reStructuredText
document.

The biggest win we get from this is disabling all `include::` style
directives giving us line numbers that are not offset by whatever
content the `include::` directive inserts.

This commit introduces the `get_initial_doctree` method which lives
alongside the existing `get_doctree` method on the
`RstLanguageServer`. It also updates the
`textDocument/documentSymbols` request to use the new method, yielding
more accurate results.