lsp: Implement dependency injection for `esbonio_setup` functions
To help minimise the amount of boilerplate for extensions, it's now
possible to request specific language features in the signature of an
`esbonio_setup` function. For example

```python
from esbonio.lsp.rst import RstLanguageServer
from esbonio.lsp.roles import Roles

def esbonio_setup(rst: RstLanguageServer, roles: Roles):
    ...
```

Then as the module is loaded, the requested language features will
automatically be passed into the function when it is called.

However, if a requested feature is not available, or the requested
language server's type is not compatible then this setup function will
not be called. This should ensure that the server degrades gracefully if
a user happens to disable a module's dependencies.