lsp: Enable dependency injection for remaining extension points
The majority of the extension loading logic has now been refactored into
a `load_extension` method available on the `RstLanaguageServer`. Given
an extension name and a setup function it handles the dependency
injection for and execution of the given setup function.

This commit adjusts the `_load_sphinx_extensions` and
`_load_sphinx_config` code to use this new method, enabling dependency
injection for the Sphinx based extension points.