lsp: Attempt to make it possible to debug server under test.
Now if you run the test suite with the environment variable
`USE_DEUBGPY` set `pytest_lsp` will launch the server wrapped in a
`debugpy` instance so it can (in theory) be debugged.

This also adds a launch configuration to VSCode that connects to an
existing `debugpy` instance. However, I am haven't been able to
successfully connect yet.