lsp: Use a text document's language id where possible
When identifying where a given position is (python vs rst vs docstring)
use the document's `lanuage_id` field where possible. This handles
situations where a user has configured `*.txt` files in VSCode to be
treated as rst files.

If the document's language id is not available, then we fall back to
using file extensions as we did previously