lsp: Allow `DirectiveLanguageFeatures` to suggest directive options
Since `DirectiveLanguageFeatures` are now responsible for discovering
the available directives, they are also the best suited to determine
which options a directive has.

Subclasses can now implement the `suggest_options` method to provide
this functionality.

Additionally, in order to correctly determine the correct implementation
given some name for a directive we need to be able to do the inverse of
the index -> suggest step of generating completions. So
`DirectiveLanguageFeatures` that provide directive implementations
should also implement the `get_implementation` method.