workflow: Create a dedicated VSCode PR workflow
This commit introduces a dedicated workflow to validate PRs made against
the VSCode extension. It uses the `paths` filter to select which files
should trigger it.

It also takes the opportunity to bump all the version numbers of the
actions used, hopefully fixing some deprecation warnings in the process.

The monolithic release pipeline will remain, but the vscode portion has
been updated to use the latest versions of actions (where possible) and
any testing steps have been removed.

Finally, the `npm deploy` script has been tweaked to (hopefully) reuse
the version of `vsix` package that was built for uploading to the
releases page - avoiding building the same package twice.