workflow: Create a dedicated lsp PR workflow.
The large test matrix of python versions and OSes that we use for the
language server is now broken out into a dedicated workflow. This should
greatly reduce the amount of unecessary work done by the release
pipeline.

This commit also bumps all the (LSP) action versions (where possible)
in the existing release workflow while dropping the steps that were only
required for testing.