lsp: Introduce `RoleLanguageFeatures`
Following on from the work on ``DirectiveLanguageFeatures`` a similar
refactoring has been performed for roles.

`RoleLanguageFeatures` are now responsible for all the 'smarts' behind
LSP features with the `Roles` language feature providing the framework
within which they operate.

Features can currently implement the following methods

- ``complete_targets`` called when generating role target completion
  items
- ``find_target_definitions`` used to implement goto definition for role
  targets
- ``get_implementation`` used to get the implementation of a role given
  its name
- ``index_roles`` used to tell the language server which roles exist
- ``resolve_target_link`` used to implement document links for role
  targets
- ``suggest_roles`` called when generating role completion suggestions

The existing "provider" approach has been deprecated along with a number
of methods on the base lanaguage servers all of which will be removed in
the ``v1.0`` release