code: **Experimental** support for  `esbonio` in vscode.dev
It is now possible to run the Sphinx based language server in
vscode.dev!

It's a bit of a hack, the biggest blocker was not being able to access
the files in the workspace within pyodide. This works around it by
having the VSCode extension copy the contents of the entire workspace
into pyodide's in-memory filesystem.

Obviously it's not ideal, and there's plenty that doesn't work still but
it works enough to let us start flushing out places within the language
server where we assume a filesystem.

Improvements to the runtime (such as using WASI?) can now be developed
in parallel.