lsp: Scaffold the `sphinx_agent`
The sphinx agent will be launched as a subprocess and wrap a Sphinx
application object, exposing a JSON-RPC API allowing the parent
process to remotely inspect and control it.

The agent process will be launched from the python interpreter used to
build the user's documentation with the agent itself being injected
into the environment with the `PYTHONPATH` env variable.

For this reason and to maximise compatibility, the agent itself can
only depend on the standard library and sphinx. This does mean having
to re-invent some of the JSON-RPC machinery, but hopefully `asyncio`
and `dataclasses` should make that relatively straightforward.

Fingers crossed it all works out!