code: Cleanup old options, add `workspace/configuration` middleware
It's looking like a lot of the old configuration options won't be
needed so they are being removed for now.

The server needs to know which python interpreter it should use when
spinning up new Sphinx instances. It would be good for the extension
to default to the Python interpreter for the current workspace -
unless told otherwise.

Unfortunately, the Python extension no longer records this as a
standard configuration option.

But adding some middleware to the language client, we can intercept
`workspace/configuration` requests from the server and inject the
Python interpreter as configured by the python extension if necessary!