lsp: Basic multi-project support!
Now when creating a client, the client is stored on the
`SphinxManager` class in a dictionary, using the Sphinx app's
`srcdir` (converted to uri) as the key.

That way, when using `get_client` to obtain the correct client for the
given uri all we have to do is check if its key is a prefix of the
given uri.

I can think of plenty of edge cases where this wouldn't necessarily
work well, but this *probably* is good enough for the 80% use case?