lsp: Proof of concept `PreviewManager` feature
Rather than spin up the http server in a separate process (like
`SphinxLanguageServer` does), it looks like we can get away with
pushing it into the server's `ThreadPoolExecutor` which should make
communication a little easier.

The idea here is that the client calls the
`esbonio.server.previewFile` command with the uri it wishes to
preview. The `PreviewManager` then spins up the http server and
otherwise configures it to serve the relevant Sphinx project.

It then replies to the client with the full uri required to view the
relevant HTML file - removing the need for each client to figure it
out themselves.

While multi-project support is possible - it's only one at a time at
the moment