sphinx_agent: Inject `webview.js` into html builds.
When the `enable_sync_scrolling` flag is set, the sphinx agent will
now inject an additional `webview.js` script into pages built with
sphinx.

This file implements a miniature JSON-RPC client over websockets that
will allow the main language server to connect to the webview and
control it.

Currently only a `view/reload` notification from the server is
supported, which, as the name suggests will allow the server to ask
the page to reload itself.

Taking this approach not only allows us to support fancy preview
features in editors that do not come with a native webview
component (like Neovim) but we get to work around a long standing
bug (#440).

This is because the page is reloading itself, we don't encounter the
hard reload required when reloading externally when crossing the
`<iframe>` barrier