lsp: Add `Uri` class
Personally, I find correctly handling filepaths and uris on both
Windows and *nix systems the hardest to get right. This class is an
attempt as resolving this once and for all.

Using the test cases[1] from VSCode's implementation of URIs as a
guide the new `Uri` class, among other things, should make it easy to
obtain consistent, canonical representations for files and uris.

This implementation also implements a few "dunder" methods so that
- paths can be joined using `/` similar to `pathlib.Path` objects
- instances are `os.PathLike`, so uris representing files on disk
  should "just work" as you'd expect with functions like
  `os.path.join`, `open` etc.

Finally, instances of the class are frozen meaning they are hashable
and can be used as dictionary keys.

[1]: https://raw.githubusercontent.com/microsoft/vscode/18136ed6ec4bfabc58664c8b656839709ea997e8/src/vs/base/test/common/uri.test.ts