lsp: Decouple the `SphinxManager` feature from the `SphinxClient`
The `SphinxManager` feature is now initialized with a factory function
that it can call to obtain a new `SphinxClient` instance. Not only
should this enable to easily write unit tests for it, but it should
make it possible for us to provide an in-process implementation for
the WASM version of the server.

To this end `SphinxClient` is now a `typing.Protocol` that specifies
the API the `SphinxManager` can rely on. What was previously called
`SphinxClient` is now the `SubprocessSphinxClient`