lsp: Add `get_user_config` method
This adds a standard `get_user_config` method to the
`EsbonioLanguageServer` that language features can use to obtain
configuration values from the client.

Currently it uses only the `workspace/configuration` request however,
at some point this should also fallback to `initializationOptions` if
necessary so that we can be compatible with any clients that don't
support the `workspace/configuration` request.

This method also handles parsing the config as a particular type.