code: Preview enhancements and fixes
Rather than recreate the HTML page each time, it's now loaded once
when the webview panel is created. To change page, the extension now
uses `postMessage()` to send a message into the wrapper webpage which
in turn updates the `src` on the embedded `<iframe>`

An indeterminate progress indicator is displayed each time `src` of
the `<iframe>` is updated. The wrapper webpage listens for a ready
message from `webview.js` after which it will remove the indicator.

The following improvements have also been implemented
- The preview window no longer reopens itself after it has been closed
- The preview window no longer steals focus
- The preview is no longer pointlessly refreshed when re-focusing the
  editor for the currently shown document.
- Thanks to improvements in the language server, the extension no
  longer attempts to show files that are not part of the Sphinx
  project.