code: Set `retainContextWhenHidden` when creating the preview pane
As the name suggests, this prevents VSCode from destroying the
contents when the preview pane is hidden.

The docs discourage you from enabling this due to it being more
resource intensive but I think the experience makes up for that as
the preview is even able to main sync scroll state while still in the
background and recovering from a reveal is near instant.

If the resource usage ever becomes an issue, we can expose an option
and fall back to the previous commit's approach if the user desires