sphinx-agent: Introduce `SphinxHandler`
The adjusts the architecture of the sphinx agent slightly.
By having the handlers attached to a parent class, it should be easier
to share state between them.

This also remove the need for an explicit `@handler` decorator as the
class is able to inspect itself as part of its `__init__` method