lsp: Resolve uris and filepaths recevied from the client
On OSes like Fedora Silverblue, `/home` is symlinked to `/var/home`.
This causes issues, since, depending on the origin of any path given
to us `/home` may or may not be the true location of the file - which
introduces consistency

By always dealing with resolved filepaths/uris we should ensure that
the language server/sphinx agent are always consistent with how they
refer to files.

This also improves the accuracy of the `get_client` method on the
`SphinxManager` by using the client's `build_file_map` (when possible)
to determine if a file is part of project