lsp: Add dedicated configuration component
The idea is that all configuration will be managed in a centralised
location. The `WorkspaceConfiguration` component will be attached to the
main language server will be responsible from merging configuration
values from a number of sources.

- Initialization options
- ``workspace/configuration`` requests
- Configuration files

While this is not implemented yet, this component will also make sure
that the configuration values are kept up to date. Eventually features
will be able to subscribe to changes in the sections that they are
interested in.