sphinx-agent: Add simple database abstraction
This adds our own custom `Sphinx` application object with just a
single extra field - a `Database` object.

The `Database` object will provide a common interface for writing data
to be shared with the parent language server. So that data from
previous sessions can be reused, the tables are only dropped when
`ensure_table` detect that the schema has changed.

SQLite does not seem to like the `?` syntax for these table
manipulation statements so at the moment the statements are built just
by concatenating strings... that's probably ok for this use case?
It's not like there's any sensitive data to leak??