sphinx-agent: Create a database during application startup
The `handlers.py` module is now a directory in which we can place
custom Sphinx extensions to extract all the data we need to power
the language server. We should be able to rely on Sphinx's native
event handling system to hook into what we need.

The first extension populates the `esbonio` field of our new subclass
of `Sphinx` with an instance of the `Database` abstraction and listens
for the `build-finished` event and populate a table with the
equivalent `build_file_map` data we used to send with the
`BuildResult` object.