lsp: Replace `build_file_map` with data from db.
Instead of relying on the sphinx-agent to broadcast the
`build_file_map` to the parent language server on each build, the
language server reads the required data from the new database instead.

This does introduce a few issues e.g. the preview no longer
automatically triggers a build at the moment, but hopefully they can
all be fixed in the long run.