lsp: Exclude certain `pytest-asyncio` releases
There appears to be some issues with the `0.23.*` release that affects
how we write tests. Most obvious is that redefining the `event_loop`
fixture is deprecated and will eventually become an error.

However, these releases also cause the test suite to hang/exit with a
"Event loop closed" error. It looks like `0.23.3` *may* fix this for
us, so let's wait for that to be released before revisiting this.

By excluding the releases instead of setting an upper bound, CI should
automatically pull in the new release when it's available and serve as
a reminder.