sphinx-agent: Add symbol support for MyST directives
It turns out the information available to rst and myst directives is,
annoyingly, subtly different.

So rather than trying to parse the directive from the document's raw
text (something not immediately accessible to MyST directives), we
rely on the information available on the directive itself.

This commit also updates our `disable_roles_and_directives` dispatcher
to lookup the real implementation of the directive so that we can
determine which options are valid.

This isn't perfect by any means
- symbol ranges are even less accurate than before
- invalid options currently prevent a directive from being recognized
  as a symbol.

But this is probably good enough for a beta :)